changes based on PR comments
